% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenFcns.R
\name{Ben.ps}
\alias{Ben.ps}
\title{Generate Benford probabilities}
\usage{
Ben.ps(digits)
}
\arguments{
\item{digits}{A significant digits vector, e.g. 1:9, 10:99}
}
\description{
\code{Ben.ps} returns a vector of Benford probabilities corresponding to the
digits of the argument vector \code{digits}.
}
\examples{
Ben.ps(1:9)

}
