% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenFcns.R
\name{bLoga}
\alias{bLoga}
\title{bLoga}
\usage{
bLoga(b, a)
}
\arguments{
\item{b}{A non-negative number.}

\item{a}{A non-negative number.}
}
\description{
\code{bLoga} is used in \code{\link{LR.genben}}.
}
\keyword{internal}
