% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenFcns.R
\name{Chi.approx}
\alias{Chi.approx}
\title{Pearson's three-moment Chi-square approximation}
\usage{
Chi.approx(x, eigvals)
}
\arguments{
\item{x}{MLDEFINITION}

\item{eigvals}{MLDEFINITION}
}
\description{
\code{Chi.approx} computes upper tail probabilities of asymptotic
distributions for the CVM type statistics using the Pearson chi-square
approximation.
}
\details{
The Chi-square approximation only requires the first three cumulants of the
statistic of interest, and is therefore faster than using the Imhof method.
See Lesperance et. al for further details.
}
\references{
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
  Wong, S. (2010)  Testing Benford's Law with the first two significant
  digits.  University of Victoria, Master's thesis.
}
\seealso{
\code{\link{Imhof}}
}
