% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenFcns.R
\name{CVMEigen}
\alias{CVMEigen}
\title{Compute eigenvalues required for function CVMStats}
\usage{
CVMEigen(digits)
}
\arguments{
\item{digits}{A significant digits vector.}
}
\value{
The output is a list of three component vectors of eigenvalues:
  \code{$eig.W} as Cramer-von Mises, \code{$eig.U} as Watson, and
  \code{$eig.A} as Anderson-Darling respectively.
}
\description{
\code{CVMEigen} computes eigenvalues needed for computation of significance
levels of Cramer-von Mises type statistics for testing \eqn{H_0:} Benford vs
\eqn{H_1:} multinomial p's.
}
\references{
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
  Wong, S. (2010)  Testing Benford's Law with the first two significant
  digits.  University of Victoria, Master's thesis.
}
\seealso{
\code{\link{CVMStats}}
}
