% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenFcns.R
\name{CVMStats}
\alias{CVMStats}
\title{Compute Cramer-von Mises Statistics}
\usage{
CVMStats(freq, method = c("W", "Wap", "U", "Uap", "A", "Aap", "X"),
  digits = 1:9)
}
\arguments{
\item{freq}{Vector of multinomial probabilities.}

\item{method}{Parameter to specify the statistic and the method by which to
  compute the statistic. This parameter can take on any of the following
  values, or a vector containing any combination of these values separated
  using commas:

\enumerate{
    \item 'W'
    \item 'Wap'
    \item 'U'
    \item 'Uap'
    \item 'A'
    \item 'Aap'
    \item 'X'
    }

    These parameter values correspond to the statistics and their methods
    respectively:

     \enumerate{
     \item Cramer-von Mises, Imhof
     \item Cramer-von Mises, approximation
     \item Watson, Imhof
     \item Watson, approximation
     \item Anderson-Darling, Imhof
     \item Anderson-Darling, approximation
     \item Pearson
     }

     Specifying no method will return all statistics and methods.}

\item{digits}{A significant digits vector. If unspecified, default is 1:9. \code{digits}
must match the frequencies \code{frequencies}.}
}
\value{
The output is a vector containing the user specified statistics by
  method.
}
\description{
\code{CVMStats} computes p-values for the CVM statistics for testing
\eqn{H_0:} probabilities follow Benford vs. the most general alternative,
\eqn{H_1:} probabilities are multinomial. The four statistics are \eqn{W^2}
(Cramer-von Mises), \eqn{U^2} (Watson), \eqn{A^2} (Anderson-Darling) and
\eqn{X^2} (Pearson's chi-square).
}
\details{
There are three CVM type statistics: Cramer-von Mises as
\code{W}, Watson as \code{U}, Anderson-Darling as \code{A}, as
well as Pearson's chi-square as \code{X}.

The three CVM type statistics can be computed using one of two
methods: Imhof's numerical method, or a chi-square approximation. The
argument \code{method} can be used to indicate which statistics to compute
with which method. Note that the chi-square approximation is faster to
compute than the Imhof numerical method. See Lesperance et. al (2016) for further
details.
}
\examples{
set.seed(123)
CVMStats((firstdigitsfreq(rnorm(100), 1)))
}
\references{
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
  Wong, S. (2010)  Testing Benford's Law with the first two significant
  digits.  University of Victoria, Master's thesis.
}
\seealso{
\code{\link{CVMEigen}}
}
