% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenFcns.R
\name{dec.objf}
\alias{dec.objf}
\title{Objective function for numerically finding MLEs under LR.dec, given z's}
\usage{
dec.objf(zs, freq)
}
\arguments{
\item{zs}{Numeric vector, defined as the difference in probabilities:
\eqn{z_i = p_i - p_{i+1}}}

\item{freq}{Vector of multinomial frequencies.}
}
\description{
\code{dec.objf} contains the (negative) objective function of the nonlinear
programming problem used to compute the MLEs in \code{LR.dec}.
}
\details{
\code{LR.dec} computes the  MLEs of the first significant digit probabilities
numerically under the alternative \eqn{H_1: p_1 \ge p_2 \dots \ge p_9} by
solving the nonlinear programming problem:

\eqn{ maximize \sum f_i*log(\sum z_j)}

subject to constraints, where \eqn{f_i} is the frequency of the ith
significant digit, and \eqn{z_i = p_i - p_{i+1}}. \code{dec.objf}
contains the negative of this expression in order to adhere to standard
minimization optimization procedure. See Lesperance et. al for
further details.
}
\references{
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
  Wong, S. (2010)  Testing Benford's Law with the first two significant
  digits.  University of Victoria, Master's thesis.
}
\seealso{
\code{\link{LR.dec}}
}
