% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenFcns.R
\name{dImhof}
\alias{dImhof}
\title{Integrand for the \code{Imhof} function}
\usage{
dImhof(u, xcrit, lambda)
}
\arguments{
\item{u}{MLDEFINITION evaluate the integrand at u}

\item{xcrit}{MLDEFINITION x critical P (Q>x)}

\item{lambda}{MLDEFINITION eigenvalues of A\%*\%Sigma}
}
\description{
\code{dImhof} contains the integrand used in the Imhof numerical method function.
}
\references{
Imhof, J. P. (1961). Computing the Distribution of Quadratic Forms in Normal Variables.
Biometrika 48, 419-426.  Imhof, J. P. (1962). Corrigenda: Computing the Distribution of
Quadratic Forms in Normal Variables. Biometrika 49, p. 284.
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
  Wong, S. (2010)  Testing Benford's Law with the first two significant
  digits.  University of Victoria, Master's thesis.
}
\seealso{
\code{\link{Imhof}}
}
