% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenFcns.R
\name{Div}
\alias{Div}
\title{Div}
\usage{
Div(a, b)
}
\arguments{
\item{a}{A non-negative number.}

\item{b}{A non-negative number.}
}
\description{
\code{Div} is used to prevent division by zero in \code{\link{CVMEigen}}.
}
\keyword{internal}
