% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenExtraFcns.R
\name{extractdigits}
\alias{extractdigits}
\title{Extract first digits from user data}
\usage{
extractdigits(x, numdigits = 1)
}
\arguments{
\item{x}{Numeric matrix or vector.}

\item{numdigits}{A positive integer, default is 1.}
}
\value{
The output is a vector of first \code{numdigits} for nonzero entries
  in \code{x} (taken in column-major order if \code{x} is a matrix).
}
\description{
\code{extractdigits} extracts the first nonzero \code{numdigits}
significant digits of the user's data, \code{x}.
}
\details{
"0" entries are omitted from the data.

The first n significant digits of a non-zero number are the first n non-zero
digit of the number. For example, the first two significant digits of "012"
are "12".

The first n significant digits of a number that has less than m < n digits is
given by the m-digit number concatenated with (m-n) zeroes. For example, the
first four significant digits of "1.2" are "1200".
}
\examples{
extractdigits(1:10, 1)
extractdigits(c(012, 1.2, 1, 0),2)

}
\references{
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
}
\seealso{
\code{\link{freqdigits}}

\code{\link{firstdigitsfreq}}
}
