% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimultaneousFcns.R
\name{f}
\alias{f}
\title{Function for Sison and Glaz's \code{v}}
\usage{
f(x, gamma1, gamma2)
}
\arguments{
\item{x}{MLDEFINITION}

\item{gamma1}{MLDEFINITION}

\item{gamma2}{MLDEFINITION}
}
\description{
\code{f} is a function required by \code{v}.
}
\references{
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
  Wong, S. (2010)  Testing Benford's Law with the first two significant
  digits.  University of Victoria, Master's thesis.
}
\seealso{
\code{\link{SimultConf}}

\code{\link{f}}
}
\keyword{internal}
