% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenExtraFcns.R
\name{freqdigits}
\alias{freqdigits}
\title{Count frequencies of a significant digits vector}
\usage{
freqdigits(x, numdigits = 1)
}
\arguments{
\item{x}{A significant digits vector.}

\item{numdigits}{A positive integer, default is 1.}
}
\value{
The output will always be a vector of length
  \eqn{9*(10^(numdigits-1))}.
}
\description{
\code{freqdigits} computes the frequencies of numbers with \code{numdigits}
first significant digits in a vector \code{x} of first digits.
}
\details{
\code{freqdigits} may return different vectors for a single \code{x}
depending upon the \code{numdigits} chosen. For example,
\code{freqdigits(1:9, 1)} will return a vector of ones of length 9, but
\code{freqdigits(1:9, 2)} will return a vector of zeroes of length 90
because the values in 1:9 all have only 1 digit.

If the vector \code{x} contains first digits of varying lengths, then \code{freqdigits}
will only count the first digits containing
\code{numdigits} digits. For example, inputting \code{x}=\{1,10,100\} with
\code{numdigits}=1 will return a vector of length 9 with a 1 in the first
position and zeroes otherwise.
}
\examples{
freqdigits(1:9, 1)
freqdigits(c(456,290,180), 3)
freqdigits(extractdigits(c(0.41, 1.25, 0.21, 0.54, 0.19),1),1)

}
\references{
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
}
\seealso{
\code{\link{extractdigits}}

\code{\link{firstdigitsfreq}}
}
