% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenFcns.R
\name{genben.ml.objf}
\alias{genben.ml.objf}
\title{Objective function for numerically finding MLE under LR.genben}
\usage{
genben.ml.objf(alpha, freq, digits)
}
\arguments{
\item{alpha}{A real number.}

\item{freq}{Vector of multinomial frequencies.}

\item{digits}{A significant digits vector.}
}
\description{
\code{genben.ml.objf} contains the (negative) objective function of the
nonlinear programming problem used to compute the MLE in \code{LR.genben}.
}
\details{
\code{LR.genben} computes the  MLE for \eqn{\alpha} numerically under the
generalized Benford alternative by maximizing a log-likelihood objective
function, or equivalently minimizing the negative of said function. See
Lesperance et. al for further details.
}
\references{
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
  Wong, S. (2010)  Testing Benford's Law with the first two significant
  digits.  University of Victoria, Master's thesis.
}
\seealso{
\code{\link{LR.genben}}

\code{\link{genben.ps}}
}
