% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenFcns.R
\name{genben.ps}
\alias{genben.ps}
\title{Generate Generalized Benford probabilities}
\usage{
genben.ps(alpha, digits = 1:9)
}
\arguments{
\item{alpha}{A real number.}

\item{digits}{A significant digits vector, default is 1:9.}
}
\description{
\code{genben.ps} returns a vector of generalized Benford probabilities
corresponding to the digits of the argument vector \code{digits}, and
the chosen \eqn{\alpha}.
}
\examples{
genben.ps(1, 1:9)

}
\references{
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
  Wong, S. (2010)  Testing Benford's Law with the first two significant
  digits.  University of Victoria, Master's thesis.
}
\seealso{
\code{\link{Ben.ps}}

\code{\link{LR.genben}}
}
