% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenFcns.R
\name{Imhof}
\alias{Imhof}
\title{Imhof numerical method}
\usage{
Imhof(x, eigvals, UPPER = NA, subdiv = 100, eps1 = 1e-04)
}
\arguments{
\item{x}{MLDEFINITION}

\item{eigvals}{MLDEFINITION}

\item{UPPER}{MLDEFINITION. If unspecified, default is NA.}

\item{subdiv}{MLDEFINITION. If unspecified, default is 100.}

\item{eps1}{MLDEFINITION. If unspecified, default is 0.0001.}
}
\description{
\code{Imhof} computes upper tail probabilities of asymptotic distributions
for the CVM type statistics using Imhof's numerical method.
}
\references{
Imhof, J. P. (1961). Computing the Distribution of Quadratic Forms in Normal Variables.
Biometrika 48, 419-426.  Imhof, J. P. (1962). Corrigenda: Computing the Distribution of
Quadratic Forms in Normal Variables. Biometrika 49, p. 284.
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
  Wong, S. (2010)  Testing Benford's Law with the first two significant
  digits.  University of Victoria, Master's thesis.
}
\seealso{
\code{\link{Chi.approx}}
}
