% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenFcns.R
\name{LR.mat.mult}
\alias{LR.mat.mult}
\title{Likelihood ratio test: Benford vs. General Multinomial}
\usage{
LR.mat.mult(freq, digits = 1:9)
}
\arguments{
\item{freq}{Vector or matrix with multinomial samples in the rows.}

\item{digits}{A significant digits vector. If unspecified, default is 1:9. \code{digits}
must match the frequencies in the rows of \code{freq}.}
}
\value{
The output is a vector of p-values, one per sample (row).
}
\description{
\code{LR.mat.mult} is a likelihood ratio test for whether the first
significant digits of a set of samples are compatible with Benford's Law,
where \eqn{H_0:} cell probabilities follow Benford's, vs. \eqn{H_1}: the most
general multinomial, i.e. all probabilities non-negative,
sum to 1.
}
\details{
The LR statistic \eqn{latexascii}{\lambda} for testing \eqn{H_0} vs.
\eqn{H_1} is:

\eqn{latexascii}{-2log\lambda = -2 \sum f_i{log(p_i) - log(f_i/n)}}

where \eqn{f_i} is the frequency of the ith first digit, and \eqn{p_i} is the
cell probability of the ith first digit.
}
\examples{
LR.mat.mult(firstdigitsfreq(c(0.41, 1.25, 0.21, 0.54, 0.19), 1))
set.seed(123)
LR.mat.mult(firstdigitsfreq(rnorm(100), 1))

}
\references{
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
  Wong, S. (2010)  Testing Benford's Law with the first two significant
  digits.  University of Victoria, Master's thesis.
}
