% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenFcns.R
\name{LR.rod}
\alias{LR.rod}
\title{Likelihood ratio test: Benford vs. Rodriguez family}
\usage{
LR.rod(frequencies, digits = 1:9)
}
\arguments{
\item{frequencies}{Vector of multinomial frequencies.}

\item{digits}{A significant digits vector. If unspecified, default is 1:9. \code{digits}
must match the frequencies \code{frequencies}.}
}
\value{
The output is a list containing a p-value \code{$SL} and a
  convergence number \code{$conv} (0 meaning successful convergence)
  respectively.

  The convergence number indicates convergence of the numerical method
  used to obtain the MLE of \eqn{\alpha}.
}
\description{
\code{LR.genben} tests whether the first significant digits frequencies of a
sample are compatible with Benford's Law using a likelihood ratio test, where
\eqn{H_0:} cell probabilities follow Benford's, vs. \eqn{H_1}:
cell probabilities follow a Rodriguez's generalization of Benford.
}
\details{
The Rodriguez formula depends upon parameter \eqn{\beta}. Because the
Rodriguez formula reduces to Benford's law when \eqn{\beta = -1}, equivalent
hypotheses for this test are \eqn{latexascii}{H_0: \beta = -1} vs.
\eqn{H_1: \beta =/= -1}.
}
\examples{
set.seed(123)
LR.rod((firstdigitsfreq(rnorm(100), 1)))

}
\references{
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
  Wong, S. (2010)  Testing Benford's Law with the first two significant
  digits.  University of Victoria, Master's thesis.
}
\seealso{
\code{\link{rodben.ps}}
}
