% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimultaneousFcns.R
\name{plotCI}
\alias{plotCI}
\title{Plot Simultaneous Confidence Intervals}
\usage{
plotCI(Lower, Upper, digits = 1:9, benps = TRUE, legloc = FALSE)
}
\arguments{
\item{Lower}{Matrix of lower interval limits.}

\item{Upper}{Matrix of upper interval limits.}

\item{digits}{A significant digits vector. If unspecified, default is 1:9.}

\item{benps}{\code{TRUE} will plot Benford probabilities on the graphs produced.}

\item{legloc}{\code{TRUE} will allow user to click to locate the legend.
If unspecified, default is FALSE. Note that this feature is only supported on
machines where \code{locator} is supported.}
}
\description{
\code{plotCI} plots the simultaneous confidence intervals from the Lower and
Upper bound matrices generated by \code{\link{SimultConf}}.
}
\details{
The Lower and Upper matrices are matrices of lower and upper limits, where
the columns represent the different interval types. Both
matrices must be of the same size.
}
\examples{
myN = c(28, 19, 18,  7,  9,  6,  3,  4,  6)
myCI = SimultConf(myN, c('Good','Fitz'))
plotCI(myCI$Lower, myCI$Upper)

}
\references{
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
(2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
simultaneous confidence intervals. PLoS one; 11(3).
Wong, S. (2010)  Testing Benford's Law with the first two significant
digits.  University of Victoria, Master's thesis.
}
\seealso{
\code{\link{SimultConf}}

\code{\link{SimultBenp}}

\code{\link{Ben.ps}}
}
