% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenFcns.R
\name{rev.cumsum}
\alias{rev.cumsum}
\title{Compute reverse cumulative sum}
\usage{
\method{rev}{cumsum}(v)
}
\arguments{
\item{v}{Numeric vector.}
}
\description{
\code{rev.cumsum} is used in \code{dec.objf}.
}
\seealso{
\code{\link{dec.objf}}
}
\keyword{internal}
