% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenFcns.R
\name{rodben.ps}
\alias{rodben.ps}
\title{Generate Rodriguez probabilities}
\usage{
rodben.ps(beta, digits)
}
\arguments{
\item{beta}{A real number.}

\item{digits}{A significant digits vector.}
}
\description{
\code{rodben.ps} returns a vector of Rodriguez family probabilities
corresponding to the digits of the argument vector \code{digits}, and
the chosen \eqn{\beta}.
}
\references{
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
  Wong, S. (2010)  Testing Benford's Law with the first two significant
  digits.  University of Victoria, Master's thesis.
}
\seealso{
\code{\link{LR.rod}}
}
