% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenFcns.R
\name{SimultBenp}
\alias{SimultBenp}
\title{Compare Simultaneous Confidence Intervals against Benford}
\usage{
SimultBenp(freq, alpha = 0.05, digits = 1:9)
}
\arguments{
\item{freq}{Vector of multinomial frequencies.}

\item{alpha}{Choose \code{alpha} to generate \eqn{(1-\alpha)}% CIs. If
unspecified, default is 0.05.}

\item{digits}{A significant digits vector which must match the frequecies \code{freq}.
If unspecified, default is 1:9.

#' @examples
set.seed(123)
SimultConf((firstdigitsfreq(rnorm(100), 1)), alpha=.05)
SimultBenp((firstdigitsfreq(rnorm(100), 1)), alpha=.05)}
}
\value{
The output is a vector containing 0's and 1's, where each 0 or 1
  corresponds to one of the seven simultaneous confidence intervals
  (Quesenberry and Hurst, Goodman, etc.). A zero indicates that at least
  Benford probability falls outside a given CI. A one indicates that all
  Benford probabilities fall within a given CI.
}
\description{
\code{SimultBenp} indicates whether the simultaneous confidence intervals
(generated by the user's significant digit frequencies in
\code{\link{SimultConf}}) cover the Benford probabilities.
}
\references{
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
  Wong, S. (2010)  Testing Benford's Law with the first two significant
  digits.  University of Victoria, Master's thesis.
}
\seealso{
\code{\link{SimultConf}}
}
