% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimultaneousFcns.R
\name{SimultConf}
\alias{SimultConf}
\title{Generate Simultaneous Confidence Intervals}
\usage{
SimultConf(N, method = c("Quesenberry", "Goodman", "BaileyAng",
  "Baileysqrt", "Fitzpatrick", "Sison", "binomial"), alpha = 0.05)
}
\arguments{
\item{N}{Vector of observed frequencies.}

\item{method}{The methods for simultaneous confidence intervals for
 multinomial proportions. This parameter can take on the following values, or
 a vector containing any combination of these values separated using commas:

 \enumerate{
      \item 'Quesenberry'
      \item 'Goodman'
      \item 'BaileyAng'
      \item 'Baileysqrt'
      \item 'Fitzpatrick'
      \item 'Sison'
      \item 'binomial'}

 Specifying no method will return all confidence intervals.}

\item{alpha}{Choose \code{alpha} to generate \eqn{(1-\alpha)}\% CIs. If
unspecified, default is 0.05.}
}
\value{
The output is two matrices: a Lower matrix containing the lower
  bounds, and an Upper matrix containing the upper bounds of the
  simultaneous confidence intervals using the methods specified by the user.
  The univariate approximate binomial CIs are always included.
}
\description{
\code{SimultConf} takes a vector of observed cell frequencies from a
multinomial distribution, \code{N}, and given significance level \eqn{\alpha},
returns simultaneous confidence intervals for proportions
using seven different methods:
\enumerate{
    \item Quesenberry and Hurst
    \item Goodman
    \item Bailey angular transformation
    \item Bailey square root transformation
    \item Fitzpatrick and Scott
    \item Sison and Glaz
    \item Univariate approximate Binomial confidence intervals.
}
}
\details{
An easy way to check whether your simultaneous confidence intervals generated
by \code{SimultConf} cover the Benford probabilities is to use
\code{\link{SimultBenp}}.

See Lesperance et. al (2016) for the formulae of each method.
}
\examples{
myCI = SimultConf(1:9, c('Good', 'Quesenberry'))
lower = myCI$Lower

}
\references{
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
  Wong, S. (2010)  Testing Benford's Law with the first two significant
  digits.  University of Victoria, Master's thesis.
}
\seealso{
\code{\link{SimultBenp}}
}
