% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimultaneousFcns.R
\name{v}
\alias{v}
\title{Function for Sison and Glaz's method}
\usage{
v(N, tau)
}
\arguments{
\item{N}{MLDEFINITION}

\item{tau}{MLDEFINITION An integer.}
}
\description{
\code{v} is a function required by SimultConf to compute Sison and Glaz's
method.
}
\references{
Lesperance M, Reed WJ, Stephens MA, Tsao C, Wilton B
  (2016) Assessing conformance with Benford's Law: goodness-of-fit tests and
  simultaneous confidence intervals. PLoS one; 11(3).
  Wong, S. (2010)  Testing Benford's Law with the first two significant
  digits.  University of Victoria, Master's thesis.
}
\seealso{
\code{\link{SimultConf}}

\code{\link{f}}
}
\keyword{internal}
